requirements.txt
aiogram==3.13.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
playwright==1.49.0
apscheduler==3.10.4
